NOISSUE Rename finetune and evaluator modules
